requirements.txt
llama-index
google-generativeai
llama-index-llms-gemini
pypdf
streamlit
llama-index-embeddings-gemini
python-dotenv
Ipython

-e .